requirements.txt
pytest
pytest-playwright
pytest-html
locust
py
jsonschema
eyes-playwright